core: fix new sampler defaults and add env variables to configure
